Improve the performance to build the SQL CLI image from 6 min to 1
