+++ sql/include/dockerfile_content.go
@@ -5,11 +5,9 @@ import "strings"
 // Dockerfile is the Dockerfile template
 
 var Dockerfile = strings.TrimSpace(`
-FROM python:%s-slim-bullseye
+FROM %s
 
 ENV ASTRO_CLI Yes
-
-RUN apt-get update -y && apt-get install -y gcc g++ libpq-dev python-dev libssl-dev libffi-dev
 
 RUN pip install astro-sql-cli==%s
 `)
